requirements.txt
uvicorn==0.30.1
fastapi==0.111.0
pydantic==2.7.4
SQLAlchemy==2.0.30
aiomysql==0.2.0
python-dotenv==1.0.1
alembic==1.7.4
